refactor: more endpoints refactored
