--- backend/api/order/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["OrderRequest"]
-
-from .order_request import OrderRequest
