--- backend/api/playout/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["PlayoutRequest"]
-
-from .playout_request import PlayoutRequest
