+++ backend/api/playout_control/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["PlayoutControl"]
+
+from .playout_control import PlayoutControl
