+++ backend/api/scheduling_templates/_utils.py
