+++ backend/api/server_info/_client_settings.py
