+++ backend/mise.toml
@@ -1,0 +1,2 @@
+[tools]
+lazydocker = "latest"
